requirements.txt
gradescope-utils>=0.3.1
libclang==18.1.1
pre-commit>=4.1.0
